// interfaces/species.ts
interface Species {
  name: string;
  classification: string;
  designation: string;
  average_height: string;
  average_lifespan: string;
  eye_colors: string;
  hair_colors: string;
  skin_colors: string;
  language: string;
  homeworld: string | null;
  people: string[];
  films: string[];
  url: string;
}

export type { Species };